Extract <BlockContent/>
